* TreeView.cs 	* TreeNode.cs 	* TreeNodeCollection.cs 	* win32functions.cs : more implementation added 	* button.cs : fix method signature
svn path=/trunk/mcs/; revision=17245

+++ mcs/class/System.Windows.Forms/System.Windows.Forms/win32functions.cs
@@ -760,7 +760,7 @@ namespace System.Windows.Forms{
 
 		internal static int INDEXTOSTATEIMAGEMASK(int i) 
 		{
-			return (int)((uint)i << 12);
+			return i << 12;
 		}
 
 		internal static int STATEIMAGEMASKTOINDEX(int i)
